Add ActionBar.Tabs (without deprecation, yeah)
Except wiring up the separate fragments, which may or may not be
instantiated is really fucking complicated, so the save button does
nothing, and the restaurant list is empty.

And if you rotate the screen, _another_ fragment gets instantiated on
top of the current one, or some shit.

I think I've done my penance for the equivalent of APT 6, anyway. I plan
on just redoing the two tabs as two activities or fragments or whatever,
and wire up my "add/edit restaurant" action for reals, before
implementing whatever the ICS-approved equivalent of APT 7 is.